+++ NEWS.md
@@ -1,3 +1,16 @@
+# BiocProject 0.1
+
+## 2019-01-28
+
+## Added
+
+* methods: `is` (overwrites this method behavior just for the class `Annotated`), `.is.project`, `samples`, `config` for signiture `Annotated`
+* functions: `.insertPEP` and `BiocProject` (the workhorse of the package)
+
+## Changed
+
+* **complete concept redesign**: no `BiocProject` class. The objects returned by the custom data reading function have to be of class `Annotated` and the `PEP` is inserted as the first element of its `metadata()` list
+
 # BiocProject 0.0.4
 
 ## 2019-01-25